+++ example/lib/features/app_master_layout.dart
@@ -1,0 +1,12 @@
+import 'package:app_features/app_features.dart';
+import 'package:example/app/widgets/bottom_nav.dart';
+import 'package:example/features/home/home_feature.dart';
+
+class AppMasterLayout extends MasterLayout {
+  @override
+  List<Feature> get features => [HomeFeature()];
+
+  @override
+  BottomNavigationBuildr get bottomNavigationBar =>
+      (shellRoute) => BottomNav(shellRoute);
+}