+++ lib/app_features.dart
@@ -4,4 +4,5 @@ export 'package:go_router/go_router.dart';
 export 'src/pages/bottom_sheet_page.dart';
 export 'src/pages/dialog_page.dart';
 export 'src/app_features.dart';
+export 'src/master_layout.dart';
 export 'src/feature.dart';