+++ lib/src/feature.dart
@@ -1,11 +1,17 @@
 import 'package:app_features/app_features.dart';
+import 'package:flutter/material.dart';
 
 abstract class Feature {
   String get name;
 
   List<GoRoute> get routes;
 
+  GlobalKey<NavigatorState> get navigatorKey =>
+      GlobalKey<NavigatorState>(debugLabel: runtimeType.toString());
+
   void get dependencies => () => {};
+
+  BottomNavigationBarItem? get bottomNavigationBarItem => null;
 
   Future<T?> push<T extends Object?>({
     String? name,
@@ -25,6 +31,17 @@ abstract class Feature {
     Object? extra,
   }) =>
       AppFeatures.router.replaceNamed(name ?? this.name,
+          pathParameters: pathParameters,
+          queryParameters: queryParameters,
+          extra: extra);
+
+  void go({
+    String? name,
+    Map<String, String> pathParameters = const <String, String>{},
+    Map<String, dynamic> queryParameters = const <String, dynamic>{},
+    Object? extra,
+  }) =>
+      AppFeatures.router.goNamed(name ?? this.name,
           pathParameters: pathParameters,
           queryParameters: queryParameters,
           extra: extra);